+++ api/generate/boundingbox/interface.ts
@@ -1,0 +1,5 @@
+import { Request } from "express";
+
+export type IGeoFenceRequest = Request & {
+  geoFenceCoords: [number, number, number, number];
+};
