+++ api/generate/boundingbox/public.json
@@ -1,0 +1,3 @@
+[
+ "1"
+]
