--- api/generate/circuit_template.ts
@@ -1,13 +1,0 @@
-export default ([
-  northEastX,
-  northEastY,
-  southWestX,
-  southWestY,
-]: Array<number>) => `
-pragma circom 2.0.0;
-
-include "../../location.circom";
-
-// geofence template (public inputs)
-component main = AssertFixedLocation(${northEastX}, ${northEastY}, ${southWestX}, ${southWestY});
-`;
