--- api/generate/interface.ts
@@ -1,5 +1,0 @@
-import { Request } from "express";
-
-export type IGeoFenceRequest = Request & {
-  geoFenceCoords: [number, number, number, number];
-};
