--- public/public.json
@@ -1,3 +1,0 @@
-[
- "1"
-]
