--- public/template_input.json
@@ -1,1 +1,0 @@
-{"latitude": 1, "longitude": 1}
