--- public/verification_key.json
@@ -1,94 +1,0 @@
-{
- "protocol": "groth16",
- "curve": "bn128",
- "nPublic": 1,
- "vk_alpha_1": [
-  "5741363035302961720749615118488422664300756350848876568662196788180627190936",
-  "9819189588515968237577831633163051380087961880518634882885643897182615223516",
-  "1"
- ],
- "vk_beta_2": [
-  [
-   "9568042134971208092375927965536209655484849621445827610152490046889597529697",
-   "15616639592806013201483655387751733369299078972726014061704196578965412845544"
-  ],
-  [
-   "8476423928477495853532342525190166252765851489626437428852497251097874163045",
-   "59641946488490647306901075611672426398756622126350287286303375996496983837"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_gamma_2": [
-  [
-   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-  ],
-  [
-   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_delta_2": [
-  [
-   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-  ],
-  [
-   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_alphabeta_12": [
-  [
-   [
-    "8967520636304234481216673301495180907329683387793866658384925333001344612045",
-    "9455251080819597839793449540899997872815026919954039395033171806459302906380"
-   ],
-   [
-    "16165714257113247801644689753425099483669029874803961242765692070851651742832",
-    "20146693228862308890840464191165188713149251602475249113250290230962167110076"
-   ],
-   [
-    "3938022324871494985189704988330587752623852039873088949226704836995877414622",
-    "15705662466690598811973818927239720872925855373625655055882877074674198981263"
-   ]
-  ],
-  [
-   [
-    "16955749537934478441297268399303358135769167490045035891684729791301793419874",
-    "13896941210638198928655727799773831413959238080491739224440286010057797578153"
-   ],
-   [
-    "9719181503645517002505042228885747195553893997360515958235698292541117951338",
-    "18487787290995181380559531551284193762054401110655577979363367232657020283069"
-   ],
-   [
-    "4055830423263130080727664054450478069035441209035952471134763614682543260774",
-    "16177008347990924209544373240312142634644299586322630915577018054318915860180"
-   ]
-  ]
- ],
- "IC": [
-  [
-   "88347229251892609243094715269411367245253491487234268449733796902132857134",
-   "9506918304420778665860583640987963352030766773455544540671836948678542171144",
-   "1"
-  ],
-  [
-   "6489089367501128891793266690883197388918139533232644881477799455863274422274",
-   "1722159477309017190241718578733355458394679984753925275241303097721046715897",
-   "1"
-  ]
- ]
-}
